+++ jupiter-rpc/src/main/java/org/jupiter/rpc/consumer/future/DefaultInvokeFuture.java
@@ -53,8 +53,13 @@ public class DefaultInvokeFuture<V> extends AbstractListenableFuture<V> implemen
 
     private static final long DEFAULT_TIMEOUT_NANOSECONDS = TimeUnit.MILLISECONDS.toNanos(JConstants.DEFAULT_TIMEOUT);
 
-    private static final ConcurrentMap<Long, DefaultInvokeFuture<?>> roundFutures = Maps.newConcurrentMapLong();
-    private static final ConcurrentMap<String, DefaultInvokeFuture<?>> broadcastFutures = Maps.newConcurrentMap();
+    private static final int FUTURES_CONTAINER_INITIAL_CAPACITY =
+            SystemPropertyUtil.getInt("jupiter.rpc.invoke.futures_container_initial_capacity", 1024);
+
+    private static final ConcurrentMap<Long, DefaultInvokeFuture<?>> roundFutures =
+            Maps.newConcurrentMapLong(FUTURES_CONTAINER_INITIAL_CAPACITY);
+    private static final ConcurrentMap<String, DefaultInvokeFuture<?>> broadcastFutures =
+            Maps.newConcurrentMap(FUTURES_CONTAINER_INITIAL_CAPACITY);
 
     private final long invokeId; // request.invokeId, 广播的场景可以重复
     private final JChannel channel;