+++ docs/release-notes.md
@@ -1,5 +1,9 @@
 Jupiter release notes
 ------------------------
+### 2018-06-06:
+
+- DefaultInvokeFuture#roundFutures初始容量配置, 避免频繁扩容
+
 ### 2018-04-13: version 1.2.21
 
 - 个别代码重新整理